backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
pandas
scikit-learn
joblib
python-socketio==5.11.1
python-engineio==4.9.0
